Move ipgenerator to its own pkg
This will be helpful in future commits where
we can leverage this from other places
outside the main cluster-manager.
