+++ go-controller/pkg/generator/ip/ip_generator.go
@@ -1,4 +1,4 @@
-package clustermanager
+package ip
 
 import (
 	"fmt"
@@ -8,21 +8,21 @@ import (
 	utilnet "k8s.io/utils/net"
 )
 
-// ipGenerator is used to generate an IP from the provided CIDR and the index.
+// IPGenerator is used to generate an IP from the provided CIDR and the index.
 // It is not an allocator and doesn't maintain any cache.
-type ipGenerator struct {
+type IPGenerator struct {
 	netCidr   *net.IPNet
 	netBaseIP *big.Int
 }
 
-// newIPGenerator returns an ipGenerator instance
-func newIPGenerator(subnet string) (*ipGenerator, error) {
+// NewIPGenerator returns an ipGenerator instance
+func NewIPGenerator(subnet string) (*IPGenerator, error) {
 	_, netCidr, err := net.ParseCIDR(subnet)
 	if err != nil {
 		return nil, fmt.Errorf("error parsing subnet string %s: %v", subnet, err)
 	}
 
-	return &ipGenerator{
+	return &IPGenerator{
 		netCidr:   netCidr,
 		netBaseIP: utilnet.BigForIP(netCidr.IP),
 	}, nil
@@ -33,7 +33,7 @@ func newIPGenerator(subnet string) (*ipGenerator, error) {
 // cidr. If suppose the subnet was - 100.88.0.0/16 and the specified
 // index is 10, it will return IPNet { IP : 100.88.0.10, Mask : 16}
 // Returns error if the generated IP is out of network range.
-func (ipGenerator *ipGenerator) GenerateIP(idx int) (*net.IPNet, error) {
+func (ipGenerator *IPGenerator) GenerateIP(idx int) (*net.IPNet, error) {
 	ip := utilnet.AddIPOffset(ipGenerator.netBaseIP, idx)
 	if ipGenerator.netCidr.Contains(ip) {
 		return &net.IPNet{IP: ip, Mask: ipGenerator.netCidr.Mask}, nil
